requirements.txt
bitsandbytes==0.41.0
cpm_kernels==1.0.11
deepspeed==0.13.1
faiss_gpu==1.7.2
jieba==0.42.1
modelscope==1.15.0
mteb==1.5.5
numpy==1.23.5
omegaconf==2.2.3
onnx==1.14.1
onnxruntime==1.15.1
onnxruntime_gpu==1.15.1
openai==1.35.13
packaging==24.1
pandas==2.0.3
peft==0.10.0
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.3.0
scipy==1.14.0
sentence_transformers==2.6.0
sentencepiece==0.1.99
seqeval==1.2.2
six==1.16.0
tensorflow==2.12.0
torch==2.0.1
tqdm==4.66.4
transformers==4.39.1
